allow updating source settings via http (#221)
