+++ plugin/src/main/java/com/github/topi314/lavasrc/plugin/config/LyricsSourcesConfig.java
@@ -1,19 +1,17 @@
-package com.github.topi314.lavasrc.plugin;
+package com.github.topi314.lavasrc.plugin.config;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
 
 
-@ConfigurationProperties(prefix = "plugins.lavasrc.sources")
+@ConfigurationProperties(prefix = "plugins.lavasrc.lyrics-sources")
 @Component
-public class SourcesConfig {
+public class LyricsSourcesConfig {
 
 	private boolean spotify = false;
-	private boolean appleMusic = false;
 	private boolean deezer = false;
-	private boolean yandexMusic = false;
-	private boolean floweryTTS = false;
 	private boolean youtube = false;
+	private boolean yandexMusic = false;
 	private boolean vkMusic = false;
 
 	public boolean isSpotify() {
@@ -24,14 +22,6 @@ public class SourcesConfig {
 		this.spotify = spotify;
 	}
 
-	public boolean isAppleMusic() {
-		return this.appleMusic;
-	}
-
-	public void setAppleMusic(boolean appleMusic) {
-		this.appleMusic = appleMusic;
-	}
-
 	public boolean isDeezer() {
 		return this.deezer;
 	}
@@ -40,28 +30,20 @@ public class SourcesConfig {
 		this.deezer = deezer;
 	}
 
-	public boolean isYandexMusic() {
-		return this.yandexMusic;
-	}
-
-	public void setYandexMusic(boolean yandexMusic) {
-		this.yandexMusic = yandexMusic;
-	}
-
-	public boolean isFloweryTTS() {
-		return this.floweryTTS;
-	}
-
-	public void setFloweryTTS(boolean floweryTTS) {
-		this.floweryTTS = floweryTTS;
-	}
-
 	public boolean isYoutube() {
 		return this.youtube;
 	}
 
 	public void setYoutube(boolean youtube) {
 		this.youtube = youtube;
+	}
+
+	public boolean isYandexMusic() {
+		return this.yandexMusic;
+	}
+
+	public void setYandexMusic(boolean yandexMusic) {
+		this.yandexMusic = yandexMusic;
 	}
 
 	public boolean isVkMusic() {
